requirements.txt
Django==3.1.7
djangorestframework==3.12.2
factory-boy==3.2.0
Faker==6.6.1
gunicorn==20.0.4
psycopg2-binary==2.8.6
pylama==7.7.1
pytest==6.2.2
pytest-django==4.1.0
pytz==2021.1
sqlparse==0.4.1
requests~=2.26.0